refactor(nat): Rename submodule "allocator" -> "allocation"
This submodule used to contain trait NatAllocator, but this is no longer
the case, and it doesn't contain elements directly related to the
allocator itself; instead it contains items resulting from an
allocation, namely, the return type for the allocator, and the error
type used by the allocator. To better reflect this change, rename the
submodule from "allocator" to "allocation".
