+++ nat/src/stateful/allocation.rs
